Examples showing how templates can be loaded from files.
The Yesod book shows these examples. hamlet4.hs reads a Lucius
CSS template from a text in a quasiquotation. hamlet5.hs reads
the template from a file. The Makefile allows building hamlet5
in either production or non-production mode. In non-production
mode, you can change the template without recompiling the
application, and it picks up the changes. Otherwise, the template
is processed at compile time and is baked into the application.