Add a scaffolded 'hello-world' for yesod.
I'm adding this just because a scaffolded site doesn't seem
to compile out of the box with Yesod 0.9.4.1 -- at least not
if you use the 'tiny' option without a database. The README
and the modified template 'hamlet/homepage.hamlet' contain
a few things of interest.

In particular, there seem to be some small rough spots getting
this to compile with 'cabal-dev'; the documentation spat out
by 'yesod' is mute about the '--dev' flag and which sub-commands
it really applies to.