Make yesod/Makefile work for a sibling cabal-dev directory.
At first I wanted to have a single cabal-dev directory for the entire
repository here, but I've decided against it because there is a good
chance that different directories may eventually need different
versions of yesod and other tools. On the one hand, this is a pain,
since yesod takes quite a while to install, but on the other hand it's
more in the spirit with a sandboxed build.

I had considered making symbolic links to a common directory, but that
pretty much undoes sandboxing the build. So I'll stick with the
redundant sandboxes for the time being.