Add log in and log out to the homebase home page.
All the pages here look really lame, but I'll try to fix
the looks in some separate commits.

Currently, showing the authentication ID is a bad idea, because
it's just a user identifier and so isn't really presentable.
It would be better to have a way for a user to add some
personal information, like a user name.

Note that in 'homepage.hamlet' I've put the text

 Logged in as

right in the template, instead of in the messages file. I'll
check to see how parametric messages are supposed to be
handled, and fix that.