Add a user's full name to the database, and edit it in a profile form.
This small change ties together persistence and forms -- and it took
me a while to figure some things out.

Now the 'User' type has a 'userFullName' field holding a nullable
'Text' value. If there is a value, we show 'Logged in as <V>' on
the home page. We also allow editing the value in a 'profile' form.

This illustrates the numerous things you have to do when adding
a new page and a new form:

- add the new 'fullName' attribute in the 'User' entity in
  'config/models'
- add a new route in 'config/routes', with GET and POST methods
- add a new handler module below 'Handler', in this case the
  'Handler.Profile' module
  - since this is a form, we need both GET and POST handler
    functions
- add the new handler module to the cabal file
- import the new handler module in 'Application'
- update the home page and 'Root' handler so it presents a link
  to the profile form
- add new messages as necessary

Much of this is boilerplate that surrounds adding a new handler.
In future, it might be a better idea to add stub handlers first
before making modifications to the persistent schema.

One thing I'm wondering about now is how to use 'setMessage' with
a localizable message using the I18N framework. I'll try to work
that out next.

Another thing is how to deal with showing form errors. I'm tempted
to try to devise a widget for showing form errors; I don't imagine
that the error messages are localizable, but I'll dig into that too.