Use 'renderMessage' with 'setMessage' in the profile form.
In order to user 'renderMessage', we need to have a reference
to the application and the list of client-supported languages.
The app is easy -- just call 'getYesod' in the handler -- but
it took me a while to figure out how to get the languages.

It turns out it's pretty easy too: use 'getRequest' and extract
the languages with 'reqLangs'. This really ought to be a
utility function, though I'm afraid to figure out what its
type would be :-).

The only other thing that is necessary is to call 'toHtml' on
the result.

Not knowing that the languages would be in the request shows
you how much I don't know about web app programming.