Use 'getMessageRender' in the profile POST handler.
A couple of commits back, I found that I could invoke
'renderMessage' by first getting the app and the languages
associated with the request, and I was thinking about
creating a helper of some sort. It turns out it already
exists: it's Yesod.Handler.getMessageRender. I just
happened to find it by accident while browsing the Haddocks
for that module. I've switched to using it.