Add 'header' and 'statusbar' divs in a couple of the pages.
I'd like most pages to have a visible header. There probably
ought to be a widget for this, but for the moment I've put
'header' divs in the home page and the profile form, but I've
put the style for 'header' in 'default-layout.lucius'. Perhaps
a better place to put 'global' styles would be a fixed style
sheet. Hmmm. Or maybe we need a widget for page headers.

I've given the header some rudimentary styling, and limited
the body to a width of 800 pixels just for fun. I've also
added a 'Not logged in' message when the user has yet to
log in.