Put page headers into their own widget.
Instead of repeating the hamlet markup for page headers, build a widget
for page headers, with its own hamlet and lucius files. This way, we can
take the '#header' style declarations out of 'default-layout.lucius',
and avoid treating it like a global variable. I think this fits with the
way Yesod intends us to use widgets -- widgets can be tiny, and the
smaller they are, the more likely they can be reused.