Give the log-in page a header.
The default log-in page has no header and just crams a bunch of
widgets together willy-nilly. It's possible to take control over
the log-in page by implementing your own handler for the page.
You do this by defining 'loginHandler' in the YesodAuth instance.

The handler is for the 'Auth' sub-site defined by the YesodAuth
package. The sub-site has its own message type, defined in
module 'Yesod.Auth.Messages'. I wanted to use the page header
widget in the new page, to at least get a header like the one
on the other pages. In order to do this, I had to generalize
the types in 'Widget.PageHeader'.

In order to complete the improvements in the log-in page, we
also need to be able to seize control of the styles. The
nicest way to do that is probably to define a log-in widget
where we associate an external 'lucius' file for the styles.

Unfortunately, the log-in forms on the page have widgets
contributed by the individual authentication plug-ins. In order
to style the form elements, we have to look at the source (or
the page source) and refer to the class names put on the
widgets when defining the styles. That's not ideal because the
class names could change out from under us, but it's not a
huge wart.